Rework the other transforms
